+++ arcade/context.py
@@ -21,7 +21,7 @@ from arcade.gl.texture import Texture2D
 from arcade.gl.vertex_array import Geometry
 from arcade.gl.framebuffer import Framebuffer
 from pyglet.math import Mat4
-from arcade.texture_atlas import TextureAtlas
+from arcade.texture_atlas import DefaultTextureAtlas
 from arcade.camera import Projector
 from arcade.camera.default import DefaultProjector
 
@@ -187,7 +187,7 @@ class ArcadeContext(Context):
         )
         self.atlas_geometry: Geometry = self.geometry()
 
-        self._atlas: Optional[TextureAtlas] = None
+        self._atlas: Optional[DefaultTextureAtlas] = None
         # Global labels we modify in `arcade.draw_text`.
         # These multiple labels with different configurations are stored
         self.label_cache: Dict[str, arcade.Text] = {}
@@ -227,20 +227,18 @@ class ArcadeContext(Context):
         )
 
     @property
-    def default_atlas(self) -> TextureAtlas:
+    def default_atlas(self) -> DefaultTextureAtlas:
         """
         The default texture atlas. This is created when arcade is initialized.
         All sprite lists will use use this atlas unless a different atlas
         is passed in the :py:class:`arcade.SpriteList` constructor.
-
-        :type: TextureAtlas
         """
         if not self._atlas:
             # Create the default texture atlas
             # 8192 is a safe maximum size for textures in OpenGL 3.3
             # We might want to query the max limit, but this makes it consistent
             # across all OpenGL implementations.
-            self._atlas = TextureAtlas(
+            self._atlas = DefaultTextureAtlas(
                 self.atlas_size,
                 border=2,
                 auto_resize=True,
