+++ tests/unit/atlas/test_update_texture_image_from_atlas.py
@@ -18,7 +18,7 @@ def test_update_texture_image_from_atlas():
     im = _create_image_with_rect((1, 1, 4, 4))
     im = im.transpose(Image.FLIP_TOP_BOTTOM)
 
-    atlas = arcade.TextureAtlas((256, 256), border=10)
+    atlas = arcade.DefaultTextureAtlas((256, 256), border=10)
     tex = arcade.Texture(im)
     atlas.add(tex)
